refactor(api,agent): share a session primitives base, isolate Enbox types
Reorganize the session/connect type surface so the shared shape lives in
one place and Enbox class types live in their own file.

- @enbox/agent: extract the `{ agent, did, delegateDid? }` triple into a
  new exported `AgentSessionPrimitives` type. `AgentSessionParams` now
  extends it (`& { recoveryPhrase?, identity }`), so the minimal session
  shape stops being re-declared in every downstream package.
- @enbox/api: move `EnboxAnonymousOptions`, `EnboxAnonymousApi`,
  `EnboxParams`, `EnboxSessionParams`, `EnboxSessionWrapper`,
  `EnboxConnectionInput`, `EnboxConnectOptions`, and `EnboxConnectResult`
  out of `enbox.ts` and into a dedicated `enbox-types.ts`. This mirrors
  the existing `protocol-types.ts` / `record-types.ts` /
  `repository-types.ts` convention, keeps the class file focused on
  behaviour, and makes the public type surface easier to discover.
- Derive `EnboxSessionParams` from `AgentSessionPrimitives` (now an alias)
  and `EnboxParams` from `Omit<AgentSessionPrimitives, 'did'> & {
  connectedDid: string }`. The `did` → `connectedDid` rename is the only
  intentional difference between the two; every other field flows from
  the shared base.
- `enbox.ts` shrinks ~83 lines; the public API surface is unchanged.

Test suites stay green: @enbox/agent 1521, @enbox/auth 478, @enbox/api 535.
